modular_summarization/requirements.txt
# Core LLM & inference stack
transformers==4.40.1
torch==2.2.1

# Structured-JSON enforcement for LLM outputs
lm-format-enforcer==0.10.4

# AWS SageMaker (optional remote inference)
boto3==1.34.129
python-dotenv==1.0.1

# Data wrangling & CSV helpers
pandas==2.2.2
pyarrow==19.0.1

# Progress bars (used in cleaning/utility scripts)
tqdm==4.67.1

# Testing
pytest==8.3.5
